fixing some errors with the getAllUsers route and thinking about doing some testing with jest

+++ controllers/UserController.js
@@ -2,11 +2,14 @@
 const User = require("../models/UserSchema");
 
 const authUser = async (req, res) => {
-  const { id ,email,password} = req.body;
+  // const { id ,email,password} = req.body;
   try {
-    const userAuth = User.findById(id);
-    if (userAuth && userAuth.email ) {
-     res.redirect('home')
+    const userAuth = User.findById(req.body.id);
+    if(userAuth){
+      if ( userAuth.email === req.body.email  && userAuth.password === req.body.password ) {
+       res.redirect('home')
+       res.status(200).send({status:"ok"})
+      }
     }
   } catch (error) {
       res.status(403).send({status:"error",message:`${error}`})
@@ -16,7 +19,6 @@ const authUser = async (req, res) => {
 const getAllUsers = async (req, res) => {
   try {
     const users = await User.find({});
-
     res.status(200).send({ res: users });
   } catch (error) {
     res.status(404).send({ message: `${error}` });
@@ -35,18 +37,18 @@ const getUser = async (req, res) => {
 };
 
 createUser = async (req, res) => {
-  //  const {author,title,description} = req.body;
+
+ const {email,password} = req.body;
+
+//  const user = User.exists(email)
+// if(user){
+//    res.status(405).send({status:"error","message":`User already exist`})
+//   }else{
+//   }
   try {
-    //  const task= Task.findOne({title,description})
-    //  if(task){
-    //     res.status(405).send({status:"error","message":`Task already exist`})
-    //  }else{
     const newUser = new User(req.body);
-
     await newUser.save();
-
     res.status(201).send({ status: "ok", message: `user created${newUser}` });
-    //  }
   } catch (error) {
     res.status(409).send({ status: "error", message: `${error}` });
   }
@@ -55,11 +57,16 @@ createUser = async (req, res) => {
 updateProfile = async (req, res) => {
   const { id } = req.params;
   try {
-    await User.findByIdAndUpdate(id, req.body);
-    res
-      .status(200)
-      .send({ status: "ok", message: `user with id:${id} has been updated` });
-  } catch (error) {
+    if(User.exists(id)){
+      await User.findByIdAndUpdate(id, req.body);
+      res
+        .status(200)
+        .send({ status: "ok", message: `user with id:${id} has been updated` });
+    }else{
+      res.status(404).send({ status: "error", message:"user not found"});
+    }
+  } 
+  catch (error) {
     res.status(404).send({ status: "error", message: `${error}` });
   }
 };
@@ -67,13 +74,13 @@ updateProfile = async (req, res) => {
 const deleteAccount = async (req, res) => {
   const { id } = req.params;
   try {
-    await User.deleteOne({ id });
-    res
-      .status(200)
-      .send({
-        status: `ok`,
-        message: `account with id:${id} has been deleted`,
-      });
+    if(User.exists(id)){
+      await User.deleteOne({ id });
+      res.status(200).send({
+          status: `ok`,
+          message: `account with id:${id} has been deleted`,
+        });
+    }else{res.status(404).send({status:"error",message:"account does not exist"})}
   } catch (error) {
     res.status(404).send({ message: `${error}` });
   }
